day 6

+++ src/day2.rs
@@ -1,13 +1,13 @@
 const INPUT: &'static str = include_str!("../input/2.txt");
 
-fn parse() -> impl Iterator<Item = (&'static str, i32)> {
+fn parse() -> impl Iterator<Item = (&'static str, i64)> {
     INPUT
         .lines()
         .flat_map(|s| s.split_once(' '))
         .flat_map(|(d, n)| Some((d, n.parse().ok()?)))
 }
 
-pub fn first() -> i32 {
+pub fn first() -> u64 {
     let (x, y) = parse()
         .map(|(d, n)| match d {
             "up" => (0, -n),
@@ -16,14 +16,14 @@ pub fn first() -> i32 {
         })
         .reduce(|(a, b), (x, y)| (a + x, b + y))
         .unwrap();
-    x * y
+    (x * y) as u64
 }
 
-pub fn second() -> i32 {
+pub fn second() -> u64 {
     let (x, y, _) = parse().fold((0, 0, 0), |(x, y, a), (d, n)| match d {
         "up" => (x, y, a - n),
         "down" => (x, y, a + n),
         _ => (x + n, y + a * n, a),
     });
-    x * y
+    (x * y) as u64
 }
